% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{copyAndCensorCohorts}
\alias{copyAndCensorCohorts}
\title{Copy the cohorts from the cohortStagingTable to the cohortTable and censor based on the minCellCount}
\usage{
copyAndCensorCohorts(
  connection,
  cohortDatabaseSchema,
  cohortStagingTable,
  cohortTable,
  targetIds,
  minCellCount,
  oracleTempSchema
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortStagingTable}{Name of the staging cohort table containing the cohorts before performing
any sub-grouping}

\item{cohortTable}{Name of the cohort table.}

\item{targetIds}{A vector containing the target cohort IDs consider when copying to the main table
This will copy over any subgroups associated with these target cohort IDs}

\item{minCellCount}{The minimum count of patients required when copying a cohor to the main table. Cohorts
at or below this threshold are excluded}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\description{
This function will copy the cohorts from the cohortStagingTable to the cohortTable that have a count
of subjects greater than the minCellCount.
}
