% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupConstruction.R
\name{createBulkSubgroup}
\alias{createBulkSubgroup}
\title{Create bulk subgroup cohorts}
\usage{
createBulkSubgroup(
  connection,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  targetIds,
  oracleTempSchema
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{targetIds}{A vector containing the target IDs to subgroup}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\description{
This function wraps together 2 calls: @insertRef{createBulkSubgroupFromFile} and @insertRef{createBulkSubgroupFromCohorts}.
This function assumes that cohorts are already generated and located in the cohortDatabaseSchema.cohortStagingTable
}
