% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupConstruction.R
\name{createBulkSubgroupFromCohorts}
\alias{createBulkSubgroupFromCohorts}
\title{Create subgroup cohorts by finding the intersection of target cohorts with subgroup cohorts}
\usage{
createBulkSubgroupFromCohorts(
  connection,
  cohortDatabaseSchema,
  cohortStagingTable,
  targetIds,
  oracleTempSchema
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortStagingTable}{Name of the staging cohort table containing the cohorts before performing
any sub-grouping}

\item{targetIds}{A vector containing the target IDs to subgroup}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\description{
This function creates subgroups of the target cohorts based on combinations defined in
inst/settings/targetSubgroupXref.csv

This function assumes that cohorts are already generated and located in the cohortDatabaseSchema.cohortStagingTable
}
