% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceFiles.R
\name{getCohortBasedSubgroup}
\alias{getCohortBasedSubgroup}
\title{Get the cohort subgroups from the resource file}
\usage{
getCohortBasedSubgroup()
}
\description{
Reads the settings in /inst/settings/CohortsToCreateSubgroup.csv
}
