% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceFiles.R
\name{getCohortsToCreate}
\alias{getCohortsToCreate}
\title{Get the full list of cohorts to create for the study}
\usage{
getCohortsToCreate(cohortGroups = getCohortGroups())
}
\description{
Reads inst/settings/CohortGroups.csv and creates a list of all cohorts
from the indvidual resource files. Returns a concatenated list of cohorts
}
