% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceFiles.R
\name{getCohortsToDeriveTarget}
\alias{getCohortsToDeriveTarget}
\title{Get the cohorts to derive from the resource file}
\usage{
getCohortsToDeriveTarget()
}
\description{
Reads the settings in /inst/settings/CohortsToDeriveTarget.csv
}
