% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceFiles.R
\name{getFeatures}
\alias{getFeatures}
\title{Get the cohort used for features from the resorce file}
\usage{
getFeatures()
}
\description{
Reads the settings in /inst/settings/CohortsToCreateFeature.csv
}
