% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceFiles.R
\name{getTargetSubgroupXref}
\alias{getTargetSubgroupXref}
\title{Get the list of sub-grouped target cohorts}
\usage{
getTargetSubgroupXref()
}
\description{
Reads the settings in /inst/settings/targetSubgroupXref.csv
}
