% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERA.R
\docType{package}
\name{HERACharacterization-package}
\alias{HERACharacterization}
\alias{HERACharacterization-package}
\title{HERACharacterization: Health Equity Research Assessment}
\description{
A large scale characterization of gender and racial disparites
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ohdsi-studies/HERACharacterization}
  \item Report bugs at \url{https://github.com/OHDSI/HERACharacterization/issues}
}

}
\keyword{internal}
